backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
alembic==1.14.0
PyMySQL==1.1.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
jdatetime==5.0.0
python-multipart==0.0.9
